requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cryptography==43.0.3
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.1.0
filelock==3.16.1
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.29.0
itemadapter==0.9.0
itemloaders==1.3.2
jedi==0.19.2
jiter==0.7.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
openai==1.55.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
Protego==0.3.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
selenium==4.26.1
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tldextract==5.1.3
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
wordcloud==1.9.4
wsproto==1.2.0
zope.interface==7.1.1
